requirements.txt
django==2.0.3
psycopg2
djangorestframework
